Rename triggers/stepfunction.py to triggers/step_function.py
